Trait contiguous_map::Key

pub trait Key where
    Self: Sized + Clone + Ord + Eq, {
    fn add_one(&self) -> Option<Self>;

    fn difference(&self, smaller: &Self) -> Option<usize>;

    fn add_usize(&self, num: usize) -> Option<Self>;
}

Trait that must be implemented for all key types used in a ContiguousMap.

See the blanket implementation for an alternative way of implementing this trait using the ToIndex and TryFromIndex traits.

Required methods

fn add_one(&self) -> Option<Self>

Gets the next adjacent key. Returns None if there is no adjacent key due to self being the max key.

fn difference(&self, smaller: &Self) -> Option<usize>

Gets the difference between this key and another one. Returns None if the difference does not fit in a usize.

fn add_usize(&self, num: usize) -> Option<Self>

Gets the key that is num steps after this key. Returns None if this overflows the key type.